+++ src/embedded/generator_file.rs
@@ -1,20 +1,49 @@
+use std::collections::HashMap;
 use std::fs::File;
 use std::io::BufReader;
 use std::io::Write;
 use std::path::PathBuf;
 
 use serde::Serialize;
+use serde_json::{json, Value};
 
 use crate::embedded;
 use crate::errors::*;
+use crate::event::Event;
 
 pub fn run(filename: PathBuf) -> Result<()> {
     let env = embedded::get_env().chain_err(|| "Env var processing failed")?;
+    let config = embedded::get_config(&env.config)?;
     let input = File::open(filename).chain_err(|| "Cannot open generator file")?;
     let reader = BufReader::new(input);
     let in_events = embedded::events(reader);
     let output_addr = env.output.ok_or("Missing value for SR_OUTPUT")?;
     let mut writer = embedded::output_writer(&output_addr)?;
+
+    for (_id, label) in config.labels {
+        let data: Value = serde_json::from_str(
+            &serde_json::to_string(&label).chain_err(|| "Serialization failure")?,
+        )
+        .chain_err(|| "Deserialization failure")?;
+        let mut data_m = data.as_object().unwrap().clone();
+        let hash = data_m
+            .get("hash")
+            .expect("hash")
+            .as_str()
+            .map(|s| String::from(s));
+        data_m.remove("hash");
+        let event = Event {
+            data: Some(json!(data_m)),
+            extra: HashMap::new(),
+            hash,
+            r#type: String::from("label"),
+            uri: None,
+        };
+        event
+            .serialize(&mut serde_json::Serializer::new(&mut writer))
+            .chain_err(|| "Event serialization failed")?;
+        writer.write(b"\n").chain_err(|| "Buffer write failed")?;
+    }
 
     for result in in_events {
         let mut event = result.chain_err(|| "Cannot parse line as JSON")?;